feat: moved helper functions to parser files
