+++ src/data_provider/CMakeLists.txt
@@ -134,6 +134,7 @@ elseif(CMAKE_CHECK_CENTOS5)
       "${CMAKE_SOURCE_DIR}/src/*Linux.cpp"
       "${CMAKE_SOURCE_DIR}/src/network/*Linux.cpp"
       "${CMAKE_SOURCE_DIR}/src/osinfo/sysOsParsers.cpp"
+      "${CMAKE_SOURCE_DIR}/src/packages/packageExcludeLinux.cpp"
       "${CMAKE_SOURCE_DIR}/src/packages/packageLinuxParser.cpp"
       "${CMAKE_SOURCE_DIR}/src/packages/packageLinuxParserRpmLegacy.cpp")
   add_definitions(-DLINUX_TYPE=LinuxType::LEGACY) # Partial compilation in legacy systems
@@ -142,6 +143,7 @@ elseif(CMAKE_SYSTEM_NAME STREQUAL "Linux")
       "${CMAKE_SOURCE_DIR}/src/*Linux.cpp"
       "${CMAKE_SOURCE_DIR}/src/network/*Linux.cpp"
       "${CMAKE_SOURCE_DIR}/src/osinfo/sysOsParsers.cpp"
+      "${CMAKE_SOURCE_DIR}/src/packages/packageExcludeLinux.cpp"
       "${CMAKE_SOURCE_DIR}/src/packages/packageLinux*.cpp"
       "${CMAKE_SOURCE_DIR}/src/packages/rpm*.cpp")
   add_definitions(-DLINUX_TYPE=LinuxType::STANDARD) # Standard compilation in compatible systems
