+++ src/data_provider/src/packages/packageExcludeMac.cpp
@@ -1,0 +1,14 @@
+/*
+ * Wazuh SYSINFO
+ * Copyright (C) 2015, Wazuh Inc.
+ * March 25, 2025.
+ *
+ * This program is free software; you can redistribute it
+ * and/or modify it under the terms of the GNU General Public
+ * License (version 2) as published by the FSF - Free Software
+ * Foundation.
+ */
+
+#include "packageExclude.hpp"
+
+void getExcludePackages(std::unordered_set<std::string>& excludedPaths) {};
