+++ src/data_provider/src/packages/packageLinuxDataRetriever.h
@@ -13,6 +13,7 @@
 #define _PACKAGE_LINUX_DATA_RETRIEVER_H
 
 #include <memory>
+#include <unordered_set>
 #include "filesystemHelper.h"
 #include "json.hpp"
 #include "sharedDefs.h"
@@ -32,6 +33,12 @@ void getPacmanInfo(const std::string& libPath, std::function<void(nlohmann::json
 void getRpmInfo(std::function<void(nlohmann::json&)> callback);
 
 /**
+ * @brief Get all python packages installed by rpm
+ * @param pythonPackages Set to be filled with all python packages found
+ */
+void getRpmPythonPackages(std::unordered_set<std::string>& pythonPackages);
+
+/**
  * @brief Fills a JSON object with all available rpm-related information for legacy Linux.
  * @param callback Callback to be called for every single element being found
  */
@@ -43,6 +50,12 @@ void getRpmInfoLegacy(std::function<void(nlohmann::json&)> callback);
  * @param callback Callback to be called for every single element being found
  */
 void getDpkgInfo(const std::string& libPath, std::function<void(nlohmann::json&)> callback);
+
+/**
+ * @brief Get all python packages installed by dpkg
+ * @param pythonPackages Set to be filled with all python packages found
+ */
+void getDpkgPythonPackages(std::unordered_set<std::string>& pythonPackages);
 
 /**
  * @brief Fills a JSON object with all available apk-related information
