+++ src/data_provider/src/packages/packageLinuxParserRpm.cpp
@@ -98,3 +98,44 @@ void getRpmInfo(std::function<void(nlohmann::json&)> callback)
 
 
 }
+
+void getRpmPythonPackages(std::unordered_set<std::string>& pythonPackages)
+{
+    const auto PYTHON_INFO_FILES = std::array {std::make_pair(std::regex(R"(^.*\.egg-info$)"), "/PKG-INFO"),
+                                               std::make_pair(std::regex(R"(^.*\.dist-info$)"), "/METADATA")};
+
+    const auto rawPythonPackagesList {UtilsWrapperLinux::exec(
+                                          "rpm -qa | grep -E 'python.*' | xargs -I {} rpm -ql {} | grep -E '\\.egg-info$|\\.dist-info$'")};
+
+    if (!rawPythonPackagesList.empty())
+    {
+        const auto rows {Utils::split(rawPythonPackagesList, '\n')};
+
+        for (const auto& row : rows)
+        {
+            std::smatch match;
+
+            for (const auto& [pattern, extraFile] : PYTHON_INFO_FILES)
+            {
+                if (std::regex_search(row, match, pattern))
+                {
+                    std::string baseInfoPath = match.str(0);
+
+                    if (std::filesystem::is_regular_file(baseInfoPath))
+                    {
+                        pythonPackages.insert(baseInfoPath);
+                    }
+                    else
+                    {
+                        std::string fullPath = baseInfoPath + extraFile;
+
+                        if (std::filesystem::exists(fullPath))
+                        {
+                            pythonPackages.insert(fullPath);
+                        }
+                    }
+                }
+            }
+        }
+    }
+}
