+++ src/data_provider/tests/sysInfoPackages/CMakeLists.txt
@@ -10,13 +10,8 @@ include_directories(${CMAKE_SOURCE_DIR}/../shared_modules/utils/tests/mocks)
 file(GLOB sysinfo_UNIT_TEST_SRC
     "*.cpp")
 
-if(CMAKE_SYSTEM_NAME STREQUAL "Linux")
-    file(GLOB UTILS_SRC "${CMAKE_SOURCE_DIR}/src/utilsWrapperLinux.cpp")
-endif(CMAKE_SYSTEM_NAME STREQUAL "Linux")
-
 add_executable(sysInfoPackages_unit_test
-    ${sysinfo_UNIT_TEST_SRC}
-    ${UTILS_SRC})
+    ${sysinfo_UNIT_TEST_SRC})
 
 target_link_libraries(sysInfoPackages_unit_test
     debug gtestd
@@ -28,6 +23,7 @@ target_link_libraries(sysInfoPackages_unit_test
     optimized gtest_main
     optimized gmock_main
     pthread
+    sysinfo
 )
 
 add_test(NAME sysInfoPackages_unit_test COMMAND sysInfoPackages_unit_test)
